+++ lib/dbee/query/sorter.rb
@@ -43,7 +43,9 @@ module Dbee
       end
 
       def ==(other)
-        other.key_path == key_path && other.direction == direction
+        other.instance_of?(self.class) &&
+          other.key_path == key_path &&
+          other.direction == direction
       end
       alias eql? ==