+++ spec/dbee/query/sorter_spec.rb
@@ -38,6 +38,11 @@ describe Dbee::Query::Sorter do
       expect(subject).to eq(object2)
       expect(subject).to eql(object2)
     end
+
+    it 'returns false unless comparing same object types' do
+      expect(subject).not_to eq('a.b.c')
+      expect(subject).not_to eq(nil)
+    end
   end
 
   describe 'direction helpers' do